requirements.txt
pandas
sqlalchemy
mysql-connector-python
python-dotenv
flask
flask-cors
streamlit
requests